Restructured to make flask server modular
